requirements.txt
PyYAML==6.0.2
PyAudio~=0.2.14